backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
geoalchemy2==0.14.1
shapely==2.0.2
geojson==3.1.0
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2
scipy==1.11.4
python-dotenv==1.0.0
python-multipart==0.0.6
aiofiles==23.2.1
websockets==12.0
pyproj==3.6.1
rasterio==1.3.8
rioxarray==0.15.0
xarray==2023.12.0
google-cloud-storage==2.14.0
ee==0.2
geemap==0.30.0
